requirements.txt
aenum==3.1.0
asgiref==3.3.4
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
Django==3.2.4
django-bootstrap3==15.0.0
django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-filter==2.4.0
django-js-asset==1.2.2
django-redis==5.0.0
django-simpleui==2021.8.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
idna==3.2
Jinja2==3.0.1
MarkupSafe==2.0.1
mysqlclient==2.0.3
Pillow==8.2.0
PyJWT==1.7.1
pytz==2021.1
qrcode==6.1
redis==3.5.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
twilio==6.63.0
urllib3==1.26.6
